Add files via upload
This intended to be a simple command-line interface for the Cardputer running MicroHydra. It will accept commands which are executed using Python’s exec and eval functions. Use d_print() to display on the Cardputer and print() for the console. It is still hit and miss on what is displayed where or if you will even get an output for commands. This has not been thoroughly tested at all.

Limitations:
The code is not considered safe or stable and should be used at your own risk. It directly executes user input with exec and eval.
The code does not handle errors in user input beyond catching exceptions from exec and eval. If a command causes an unhandled exception, it could crash the program.
The code does not support complex data types or operations beyond what can be handled by exec and eval.

Expected Operation: When the user types a command and presses Enter, the command is executed and the result is displayed on the screen. If the command is an assignment (it contains “=”), no result is displayed such as defining a variable i.e. (a=1) has no output. If an error occurs while executing the command, an error message is displayed.

Please note that this is a very basic implementation.